Added first waves with new creator
Added 5 generic waves for trial running.

+++ Project Files/Epitaph/Assets/Path Tool/PathFollower/PathFollow.cs
@@ -15,7 +15,7 @@ public class PathFollow : MonoBehaviour
     public bool moveToMirrored = false;
 
     public int speed = 10;
-    public int waitTime;
+    public float waitTime;
 
     int currentNavDenominator = 0;
